Add files via upload
This version of the MM model uses and array structure. A subscript structure named 'Diagnoses' was created. I has for levels,  or elements, namely: AUD, DEP, OUD, OtherNeeds. This is consistent with our discussion on Thursday, Sept 28.  Tom made this change to the model. Thanks Tom! I cannot get the model to run because it is not liking the existing ModelParameters.xlsx  file.